+++ src/main/java/terramine/common/item/magic/FallenStarItem.java
@@ -1,9 +1,0 @@
-package terramine.common.item.magic;
-
-import terramine.common.item.TerrariaItemConfigurable;
-
-public class FallenStarItem extends TerrariaItemConfigurable {
-    public FallenStarItem(Properties properties) {
-        super(properties);
-    }
-}