+++ src/main/java/terramine/common/world/DungeonGenerator.java
@@ -23,6 +23,7 @@ import net.minecraft.world.phys.shapes.BooleanOp;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.apache.commons.lang3.mutable.MutableObject;
+import terramine.TerraMine;
 
 import java.util.*;
 import java.util.function.Predicate;
@@ -30,7 +31,7 @@ import java.util.function.Predicate;
 public class DungeonGenerator {
 
     public static Optional<Structure.GenerationStub> generate(Structure.GenerationContext inContext, BlockPos pos, int size, Holder<StructureTemplatePool> startJigsawName, Holder<StructureTemplatePool> startRoomPool) {
-        if (size <= 0)
+        if (size <= 0 || TerraMine.CONFIG.worldgen.structures.disableDungeon)
             return Optional.empty();
 
         RegistryAccess registryManager = inContext.registryAccess();