+++ src/main/java/terramine/common/world/SurfaceChestFeature.java
@@ -10,6 +10,7 @@ import net.minecraft.world.level.block.entity.RandomizableContainerBlockEntity;
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.FeaturePlaceContext;
 import net.minecraft.world.level.levelgen.feature.configurations.NoneFeatureConfiguration;
+import terramine.TerraMine;
 import terramine.common.init.ModBlocks;
 import terramine.common.init.ModLootTables;
 
@@ -42,7 +43,7 @@ public class SurfaceChestFeature extends Feature<NoneFeatureConfiguration> {
 
 		generateContainer(level, positions.remove(0), random);
 
-		return true;
+		return !TerraMine.CONFIG.worldgen.caveChest.disableChests;
 	}
 
 	public boolean blockWaterCheck(BlockPos pos, WorldGenLevel level) {