+++ src/main/java/terramine/common/world/biome/BiomeAdder.java
@@ -26,20 +26,29 @@ public class BiomeAdder extends Region {
     public void addBiomes(Registry<Biome> registry, Consumer<Pair<Climate.ParameterPoint, ResourceKey<Biome>>> mapper)
     {
         if (TerraMine.CONFIG.worldgen.evilBiomeEnabled) {
-            //this.addBiomeSimilar(mapper, Biomes.PLAINS, ModBiomes.CORRUPTION);
-            //this.addBiomeSimilar(mapper, Biomes.DESERT, ModBiomes.CORRUPTION_DESERT);
-
-            if (!ModComponents.EVIL_TYPE.get(SyncedBooleanComponent.getLevelData()).get()) {
-                this.addModifiedVanillaOverworldBiomes(mapper, builder -> {
-                    builder.replaceBiome(Biomes.PLAINS, ModBiomes.CORRUPTION);
-                    builder.replaceBiome(Biomes.DESERT, ModBiomes.CORRUPTION_DESERT);
-                });
-            } else {
-                this.addModifiedVanillaOverworldBiomes(mapper, builder -> {
-                    builder.replaceBiome(Biomes.PLAINS, ModBiomes.CRIMSON);
-                    builder.replaceBiome(Biomes.DESERT, ModBiomes.CRIMSON_DESERT);
-                });
+            if (!TerraMine.CONFIG.worldgen.forceCorruption && !TerraMine.CONFIG.worldgen.forceCrimson) {
+                if (!ModComponents.EVIL_TYPE.get(SyncedBooleanComponent.getLevelData()).get()) {
+                    addCorruptionBiomes(mapper);
+                } else {
+                    addCrimsonBiomes(mapper);
+                }
+            } else if (TerraMine.CONFIG.worldgen.forceCorruption) {
+                addCorruptionBiomes(mapper);
             }
         }
+    }
+
+    private void addCorruptionBiomes(Consumer<Pair<Climate.ParameterPoint, ResourceKey<Biome>>> mapper) {
+        this.addModifiedVanillaOverworldBiomes(mapper, builder -> {
+            builder.replaceBiome(Biomes.PLAINS, ModBiomes.CORRUPTION);
+            builder.replaceBiome(Biomes.DESERT, ModBiomes.CORRUPTION_DESERT);
+        });
+    }
+
+    private void addCrimsonBiomes(Consumer<Pair<Climate.ParameterPoint, ResourceKey<Biome>>> mapper) {
+        this.addModifiedVanillaOverworldBiomes(mapper, builder -> {
+            builder.replaceBiome(Biomes.PLAINS, ModBiomes.CRIMSON);
+            builder.replaceBiome(Biomes.DESERT, ModBiomes.CRIMSON_DESERT);
+        });
     }
 }